+++ ClogAmpSwift/en.lproj/Main.strings
@@ -482,3 +482,9 @@
 "RVl-Gl-i5T.title" = "Show Beats";
 
 "z0e-vD-qbE.title" = "Open view after loading a song";
+
+"bdD-VI-P7y.title" = "Beat Change Behaviour";
+
+"u5w-Sf-Vic.title" = "Conformity factor:";
+
+"0iU-M9-9sS.title" = "Determine BPM (click)";
